Pass user state down to registration pages
This commit a pattern known as Prop Drilling in React terminology.

I wanted to abstract the User instance, `userModel`, because it lacked associated reactive functionality.
Should client code decide to make changes to userModel, those changes would not be reflected elsewhere.
In addition, `userModel` was a global instance that could be accessed anywhere,
complicating debugging of the application.

This alternative design exposes the User instance as a BehaviourSubject.
Components that support CRUD operations can subscribe to the subject and
*react* to the data source changes whenever they occur.

Since the state is now managed in the parent `Orientation` component,
I removed the update hooks in the page (child) components.