Fetch user with identity info in JWT token
Using the JWT custom claims, this commit dyanamically determines whether a user is visting Regenquest for the first time
by querying the CrossPlatform.user collection.
If it the first visit, a new user document is created on MongoDB.

Notes:
User ID could be undefined. If a user doesn't exist on Regenquest yet, there is no ID yet.
Hence we must contact the DB to replace this value. Recall the definition for `undefined` in the wiki.

Despite this revelation, ID is still considered required. If ID is set to `undefined`,
the state of the entire instance will be invalid.